Clean-up older surface_role and panel_type
This patch removes surface_role and panel_type and plug-in ShellSurface
into WebAppWayland.

Additionally we migrate PanelEdge and SurfaceType object straight in
ShellSurface as I can't seem to make them opaque such that passing a
AglShellSurface pointer to the core side as a ShellSurface works out.

Bug-AGL: SPEC-4009
