+++ src/agl/WebAppWaylandAGL.cpp
@@ -18,6 +18,6 @@
 #include "ApplicationDescription.h"
 
 WebAppWaylandAGL::WebAppWaylandAGL(const std::string& winType, std::shared_ptr<ApplicationDescription> desc)
-  : WebAppWayland(winType, desc->surfaceId(), desc->widthOverride(), desc->heightOverride(), kUndefinedDisplayId, "", desc->surfaceRole(), desc->panelType())
+  : WebAppWayland(winType, desc->surfaceId(), desc->widthOverride(), desc->heightOverride(), kUndefinedDisplayId, "")
 {
 }
